Finalised the dbus control of the ringer; refactor ringer into own module; remove ringer tests from handset button
