+++ manager.py
@@ -59,6 +59,7 @@ class PhoneManager(object):
         self._thread.start()
         self.loop_started = True
 
+
     def _listen_to_phone_ready_service(self):
         """
             Listen for the emit signal from custom service org.frank. Only necesary when no modem was
@@ -86,6 +87,9 @@ class PhoneManager(object):
             self.active_call_path = self.bt_conn.modem_object.object_path
             self._setup_volume_control()
 
+    def null_handler(self,value):
+        pass
+
     def set_call_in_progress(self, path, properties):
         """
         Event triggered when a call is initiated.
@@ -100,14 +104,20 @@ class PhoneManager(object):
         if direction == 'incoming':
             print(F"Inbound call detected on {path}")
             self.active_call_path = path
-            self.status_service.ring(config.RING_START)
+            self.status_service.send_to_ringer(config.RING_START, reply_handler=self.null_handler,
+                                               error_handler=self.null_handler)
         else:
             print("Originating outbound call")
             self.active_call_path = None
 
     def answer_call(self):
-        """ Answer the call on the modem path specified by self.active_call_path """
-        self.status_service.ring(config.RING_STOP)
+        """
+            Answer the call on the modem path specified by self.active_call_path
+        """
+
+        """ First thing is to stop the ringer via the Dbus singalling."""
+        self.status_service.send_to_ringer(config.RING_STOP, reply_handler=self.null_handler,
+                                           error_handler=self.null_handler)
         call = dbus.Interface(self.bus.get_object('org.ofono', self.active_call_path), 'org.ofono.VoiceCall')
         time.sleep(2)
         call.Answer()
@@ -119,9 +129,11 @@ class PhoneManager(object):
         :param object: The address of the call object from ofono (just as reference, cannot be fetched anymore)
         :return:
         """
-        print("Call ended!")
+        print("Call ended.")
         self.call_in_progress = False
-        self.status_service.ring(config.RING_STOP)
+        """Send the ringer_stop signal to the RingerManager to stop the ringing"""
+        self.status_service.send_to_ringer(config.RING_STOP, reply_handler=self.null_handler,
+                                           error_handler=self.null_handler)
 
     def end_call(self):
         """
